Changed the way this is working.
The CivitAI is not exclusive filtering, so if the nsfw level is set it also returns images lower than this level.

So I implemented a short search routine to ensure having the right nsfw level like set in the constructor.

Furthermore I fixed a bug with setting nsfw to None instead of 'None'